+++ trunk/source/TraceAbstraction/src/de/uni_freiburg/informatik/ultimate/plugins/generator/traceabstraction/singletracecheck/TraceCheckerStatisticsGenerator.java
@@ -102,7 +102,7 @@ public class TraceCheckerStatisticsGenerator extends StatisticsGeneratorWithStop
 		}
 	}
 	
-	private long computeLongSumOfIntArray(final int[] arr) {
+	private long computeLongSumOfIntArray(final long[] arr) {
 		long sum = 0;
 		for (int i = 0; i < arr.length; i++) {
 			sum += arr[i];
